refactor: 更新相关服务类以适配新的文章服务架构
- ArticleVersionServiceImpl 更新服务依赖，适配新的命令查询分离架构
- CommentServiceImpl 调整文章相关统计逻辑
- ReadCountService 及其实现类更新文章计数逻辑
- UserFootService 相关类更新用户足迹统计功能
- UserCacheService 优化缓存逻辑以支持新的服务架构
- 确保所有依赖文章服务的模块正确适配重构后的接口

+++ bytelogs-service/src/main/java/top/harrylei/forum/service/statistics/repository/dao/ReadCountDAO.java
@@ -69,8 +69,8 @@ public class ReadCountDAO extends ServiceImpl<ReadCountMapper, ReadCountDO> {
      * @param contentType 内容类型
      * @return 阅读量
      */
-    public Integer getReadCount(Long contentId, Integer contentType) {
+    public Long getReadCount(Long contentId, Integer contentType) {
         ReadCountDO readCount = getByContentIdAndType(contentId, contentType);
-        return readCount != null ? readCount.getCnt() : 0;
+        return readCount != null ? readCount.getCnt().longValue() : 0L;
     }
 }
